Material Fixes - Fixes the materials on the main trees and little sideways trees. - Disables fancy water to make things run faster.
