+++ Assets/BuildSettings.asset
@@ -14,16 +14,15 @@ MonoBehaviour:
   gameName: Godlands
   version:
     major: 0
-    minor: 10
+    minor: 11
     patch: 0
-    extra: 
+    extra: 1
   buildSettings:
-  - target: 2
-    options: 0
   - target: 5
     options: 0
+  - target: 2
+    options: 0
   scenes:
-  - controlsScreen
   - BlockOut
   _lastVersion:
     major: 0
