+++ Assets/Editor/BuildScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 949947933f1ae1b4fbfc85a17c0b4372
+timeCreated: 1430855851
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
