+++ Assets/Prefabs/Actors/Buddy.prefab
@@ -1523,7 +1523,7 @@ MonoBehaviour:
     min: 12
     max: 20
   _age: 1
-  _adultAge: 2
+  _adultAge: 5
   _startingHappiness: .300000012
   _happiness: 0
   _stomachRumbleSound: {fileID: 8247732, guid: fa0c1ec00694aea42a8c120250a0084d, type: 2}
