+++ Assets/Scenes/BlockOut.unity
@@ -8669,7 +8669,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -10707,7 +10707,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -13619,7 +13619,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -13681,7 +13681,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -18350,7 +18350,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -24708,16 +24708,16 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a3d3ef1a5bbfb4e0a910fbbe5830b1f9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  waterMode: 2
+  waterMode: 0
   disablePixelLights: 1
   textureSize: 256
   clipPlaneOffset: .0700000003
   reflectLayers:
     serializedVersion: 2
-    m_Bits: 4294967295
+    m_Bits: 1
   refractLayers:
     serializedVersion: 2
-    m_Bits: 4294967295
+    m_Bits: 1
 --- !u!23 &433876040
 MeshRenderer:
   m_ObjectHideFlags: 0
@@ -56532,7 +56532,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -65380,7 +65380,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -70347,7 +70347,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -95580,7 +95580,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -104301,7 +104301,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -107420,7 +107420,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -107884,7 +107884,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -108976,7 +108976,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -109689,7 +109689,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 20caf0ced1e694d9e84ee5a8cb8eb333, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -110906,7 +110906,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
@@ -113375,7 +113375,7 @@ MeshRenderer:
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_Materials:
-  - {fileID: 2100000, guid: 9622c56a07395284fb4e75bace60238a, type: 2}
+  - {fileID: 2100000, guid: 96c885a22fa574c5ab3f97cee17a0ef5, type: 2}
   m_SubsetIndices: 
   m_StaticBatchRoot: {fileID: 0}
   m_UseLightProbes: 1
